refactor: Disable `useMatrixAction`'s action when client is unavailable
